base cap_ipc_lock on file instead of file notify
To be allowed to use the mlock option when running vault with the
permission of a user other than root, such as 'vault', the
binary (/usr/local/bin/)vault needs to have the cap_ipc_lock capability
set.

Currently, the capabilities are refreshed only when the binary
changes. Failures to set the capability are reported once and ignored
afterwards.

Making the exec-resource which sets the capability depend on a
test-condition instead of a change of the binary makes puppet check the
capabilites on the file on each run and changes the capability if
cap_ipc_lock is missing on the file.

	modified:   manifests/install.pp