Enforce setcap ordering when installing from repo
Adds a dependency between package installation and setting file
capabilities when using the repo installation method. Without this,
setcap is attempted before the vault binary exists, and causes the
puppet run to fail.

Includes test that ensures ordering of capability when enabled an
package is installed via repository only.

Fixes #92