Vault config need to have proper mode
Allow customisation of the mode, and putting a secure default.
Vault configuration does need to be hidden a little bit.
When using Consul backend with proper ACLs to protect the vault/
KV path, the Consul ACL token needs to be put in the vault
configuration. File mode defaults to 0644 in puppet, which leave
the token exposed to others, allowing anyone connecting to have
read access to the vault/ path, making Consul ACLs pointless.

Calling it `config_mode` to be consistent with nomad and consul
modules that can be used along with the vault module.